# swagger.yaml
openapi: 3.0.0
info:
  title: Music Service API
  description: API для музыкального сервиса
  version: 1.0.0
servers:
  - url: /api/v1
    description: Основной API сервер
paths:
  /users:
    post:
      summary: Регистрация нового пользователя
      operationId: registerUser
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                  example: "user123"
                password:
                  type: string
                  example: "password123"
              required:
                - login
                - password
      responses:
        '201':
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Некорректные данные запроса
        '500':
          description: Ошибка сервера
  /users/auth:
    post:
      summary: Аутентификация пользователя
      operationId: authenticateUser
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                  example: "user123"
                password:
                  type: string
                  example: "password123"
              required:
                - login
                - password
      responses:
        '200':
          description: Аутентификация успешна
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  session:
                    $ref: '#/components/schemas/Session'
        '401':
          description: Неверные учетные данные
        '500':
          description: Ошибка сервера
  /users/logout:
    post:
      summary: Выход из аккаунта
      operationId: logoutUser
      tags:
        - users
      responses:
        '200':
          description: Успешный выход
        '401':
          description: Не авторизован
        '500':
          description: Ошибка сервера
  /users/{id}:
    get:
      summary: Получить профиль пользователя
      operationId: getUserProfile
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Пользователь не найден
        '500':
          description: Ошибка сервера
    
    delete:
      summary: Удалить пользователя
      operationId: deleteUser
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Пользователь успешно удален
        '404':
          description: Пользователь не найден
        '500':
          description: Ошибка сервера
  /users/{id}/permissions:
    patch:
      summary: Обновить разрешения пользователя
      operationId: updateUserPermissions
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permission:
                  type: string
                  enum: [user, admin, moderator]
                  example: "admin"
              required:
                - permission
      responses:
        '200':
          description: Разрешения обновлены
        '400':
          description: Некорректные данные запроса
        '404':
          description: Пользователь не найден
        '500':
          description: Ошибка сервера
  /tracks:
    get:
      summary: Поиск треков
      operationId: searchTracks
      tags:
        - tracks
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            example: "song title"
      responses:
        '200':
          description: Список найденных треков
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
        '400':
          description: Некорректные параметры запроса
        '500':
          description: Ошибка сервера

    post:
      summary: Загрузить новый трек (только для администраторов)
      operationId: uploadTrack
      security:
        - BearerAuth: []
      tags:
        - tracks
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: MP3 файл трека
                title:
                  type: string
                  example: "Название трека"
                artist_name:
                  type: string
                  example: "Имя исполнителя"
                duration:
                  type: integer
                  description: Длительность трека в секундах
                  example: 180
                album_id:
                  type: string
                  format: uuid
                  description: ID альбома (обязательно)
                  example: "123e4567-e89b-12d3-a456-426614174000"
                cover_url:
                  type: string
                  format: uri
                  description: URL обложки трека (необязательно)
                genre_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: |
                    Массив ID жанров для трека (опционально). Если не передан, трек создаётся без жанров.
              required:
                - file
                - title
                - artist_name
                - album_id
      responses:
        '201':
          description: Трек успешно загружен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Некорректные данные запроса
        '401':
          description: Не авторизован
        '403':
          description: Недостаточно прав (требуются права администратора)
        '500':
          description: Ошибка сервера
  /tracks/{id}:
    get:
      summary: Получить детали трека
      operationId: getTrackDetails
      tags:
        - tracks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Детали трека
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackDetails'
        '404':
          description: Трек не найден
        '500':
          description: Ошибка сервера
    
    delete:
      summary: Удалить трек (только для администраторов)
      operationId: deleteTrack
      security:
        - BearerAuth: []
      tags:
        - tracks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Трек успешно удален
        '401':
          description: Не авторизован
        '403':
          description: Недостаточно прав (требуются права администратора)
        '404':
          description: Трек не найден
        '500':
          description: Ошибка сервера
...
  /tracks/{id}/play:
    post:
      summary: Воспроизвести трек
      operationId: playTrack
      tags:
        - tracks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Трек воспроизводится
        '404':
          description: Трек не найден
        '500':
          description: Ошибка сервера
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        login:
          type: string
        permission:
          type: string
          enum: [user, admin, moderator]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - login
        - permission
        - created_at
        - updated_at
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        token:
          type: string
        expires_at:
          type: string
          format: date-time
      required:
        - id
        - token
        - expires_at
    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        duration:
          type: integer
          description: Продолжительность трека в секундах
        file_path:
          type: string
          description: Относительный путь к MP3 файлу
        album_id:
          type: string
          format: uuid
          description: ID альбома, к которому принадлежит трек (обязательно)
        artist_name:
          type: string
        cover_url:
          type: string
          description: URL обложки трека
        added_date:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        play_count:
          type: integer
          description: Количество прослушиваний
        genres:
          type: array
          items:
            $ref: '#/components/schemas/Genre'
          description: Список жанров трека
      required:
        - id
        - title
        - duration
        - file_path
        - artist_name
        - added_date
        - updated_at
    TrackDetails:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        duration:
          type: integer
          description: Продолжительность трека в секундах
        artist_name:
          type: string
        cover_url:
          type: string
          description: URL обложки трека
        added_date:
          type: string
          format: date-time
        play_count:
          type: integer
          description: Количество прослушиваний
        album_id:
          type: string
          format: uuid
          description: ID альбома, к которому принадлежит трек
        album_title:
          type: string
          description: Название альбома, если трек принадлежит альбому
        album_artist:
          type: string
          description: Исполнитель альбома, если трек принадлежит альбому
        genres:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
      required:
        - id
        - title
        - duration
        - artist_name
        - added_date
        - play_count
...
    Genre:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      required:
        - id
        - name
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 